Implement WebAssembly support check before executing content script
